# src/swagger.json
{
    "openapi": "3.0.0",
    "info":{
        "title": "API de PetShop",
        "description": "Essa API tem como objetivo o control de uma PetShop",
        "contact": {
            "email": "eduardo.alejandro.lizarraga@gmail.com"
        },
        "version": "1.0.0"
    
    },
    "servers":[
        {
            "url": "http://localhost:8080/api",
            "description": "API PetShop"
        }
    ],
    "paths": {
...
        "/pet":{
            "post": {
                "summary": "Cadastro do pet",
                "description": "Essa rota será responsável por cadastrar um novo pet",
                "tags": ["pet"],
                "requestBody": {
                    "content":{
                        "application/json":{
                            "schema": {
                                "$ref": "#/components/schemas/pet"

                            },
                            "examples":{
                                "pet":{
                                    "value": {
                                        "nroPet": 10,
                                        "nome": "marron",
                                        "tipo": "cachorro",
                                        "raça": "boxer",
                                        "idade": 2,
                                        "peso": 12,
                                        "tutorCpf": 70884720233
                                    }
                                }
                            }
                        }
                    }
                },
                "responses":{
                    "400": {
                        "description": "Erro na requisição"
                    },
                    "500": {
                        "description": "Erro no Servidor"
                    },
                    "201":{
                        "description": "criado com sucesso!",
                        "content":{
                            "application/json":{
                                "schema":{
                                    "type":"object",
                                    "items":{
                                        "$ref":"#/components/schemas/pet"
                                    }

                                }
                            }
                        }
                    }
                
                }
            }
        },
        "/pet/cpf/{cpf}":{
            "get":{
                "description": "Busca de pet pelo cpf do tutor",
                "summary": "Busca de pet pelo cpf do tutor",
                "tags":["pet"],
                "parameters": [
                    {
                        "name": "cpf",
                        "in": "path",
                        "description": "cpf do tutor",
                        "required": true
                    }
                ],
                "responses":{
                    "200":{
                        
                        "description": "OK!",
                        "content":{
                            "application/json":{
                                "schema":{
                                    "type":"object",
                                    "items":{
                                        "$ref":"#/components/schemas/pet"
                                    }

                                }
                            }
                        }
                    
                        
                        
                    },
                    "400":{
                        "description": "Erro na requisição"
                        
                    },

                    "404":{
                        "description": "não encontrado"
                    },

                    "500":{
                        "description": "Erro no Servidor"
                    }
                  

                }

            }
        },
        "/pets":{
            "get":{
                "description": "Busca de todos os pets",
                "summary": "Busca de todos os pets",
                "tags":["pet"],
                "responses":{
                    "200":{
                        
                        "description": "OK!",
                        "content":{
                            "application/json":{
                                "schema":{
                                    "type":"array",
                                    "items":{
                                        "$ref":"#/components/schemas/pet"
                                    }

                                }
                            }
                        }
                    
                        
                        
                    },
                    "400":{
                        "description": "Erro na requisição"
                        
                    },

                    "500":{
                        "description": "Erro no Servidor"
                    }

                }

            }
        },
        "/pet/{id}":{
            "get":{
                "description": "Busca de pet pelo id",
                "summary": "Busca de pet pelo id",
                "tags":["pet"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id do pet ",
                        "required": true
                    }
                ],
                "responses":{
                    "200":{
                        
                        "description": "OK!",
                        "content":{
                            "application/json":{
                                "schema":{
                                    "type":"object",
                                    "items":{
                                        "$ref":"#/components/schemas/pet"
                                    }

                                }
                            }
                        }
                    
                        
                        
                    },
                    "400":{
                        "description": "Erro na requisição"
                        
                    },

                    "404":{
                        "description": "não encontrado"
                    },

                    "500":{
                        "description": "Erro no Servidor"
                    }

                }

            },

            "put":{
                "description": "Alterar o pet pelo id",
                "summary": "Alterar o pet pelo id",
                "tags":["pet"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id do pet para alterar",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content":{
                        "application/json":{
                            "schema": {
                                "$ref": "#/components/schemas/pet"

                            },
                            "examples":{
                                "pet":{
                                    "value": {
                                        "nroPet": 10,
                                        "nome": "marron",
                                        "tipo": "cachorro",
                                        "raça": "boxer",
                                        "idade": 2,
                                        "peso": 12,
                                        "tutorCpf": 70884720233
                                    }
                                }
                            }
                        }
                    }
                },
                "responses":{
                    "200":{
                        
                        "description": "OK!",
                        "content":{
                            "application/json":{
                                "schema":{
                                    "type":"object",
                                    "items":{
                                        "$ref":"#/components/schemas/pet"
                                    }

                                }
                            }
                        }
                    
                        
                        
                    },
                    "400":{
                        "description": "Erro na requisição"
                        
                    },

                    "404":{
                        "description": "não encontrado"
                    },

                    "500":{
                        "description": "Erro no Servidor"
                    }

                }
            },
            "delete":{
                "description": "Exclui o pet pelo id",
                "summary": "Exclui o pet pelo id",
                "tags":["pet"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id do pet para Exclui",
                        "required": true
                    }
                ],
                
                "responses":{
                    "200":{
                        
                        "description": "OK!",
                        "content":{
                            "application/json":{
                                "schema":{
                                    "type":"object",
                                    "items":{
                                        "$ref":"#/components/schemas/pet"
                                    }

                                }
                            }
                        }
                    
                        
                        
                    },
                    "400":{
                        "description": "Erro na requisição"
                        
                    },

                    "404":{
                        "description": "não encontrado"
                    },

                    "500":{
                        "description": "Erro no Servidor"
                    }

                }

            }
        },
        "/pet/nroPet/{nroPet}":{
            "get":{
                "description": "Busca de pet pelo numero do pet",
                "summary": "Busca de pet pelo numero do pet",
                "tags":["pet"],
                "parameters": [
                    {
                        "name": "nroPet",
                        "in": "path",
                        "description": "numero do pet",
                        "required": true
                    }
                ],
                "responses":{
                    "200":{
                        
                        "description": "OK!",
                        "content":{
                            "application/json":{
                                "schema":{
                                    "type":"object",
                                    "items":{
                                        "$ref":"#/components/schemas/pet"
                                    }

                                }
                            }
                        }
                    
                        
                        
                    },
                    "400":{
                        "description": "Erro na requisição"
                        
                    },

                    "404":{
                        "description": "não encontrado"
                    },

                    "500":{
                        "description": "Erro no Servidor"
                    }
                  

                }

            },
            "put":{
                "description": "Alterar o pet pelo numero do pet",
                "summary": "Alterar o pet pelo numero do pet",
                "tags":["pet"],
                "parameters": [
                    {
                        "name": "nroPet",
                        "in": "path",
                        "description": "numero do pet para alterar",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content":{
                        "application/json":{
                            "schema": {
                                "$ref": "#/components/schemas/pet"

                            },
                            "examples":{
                                "pet":{
                                    "value": {
                                        
                                        "nome": "marron",
                                        "tipo": "cachorro",
                                        "raça": "boxer",
                                        "idade": 2,
                                        "peso": 12,
                                        "tutorCpf": 70884713296
                                    }
                                }
                            }
                        }
                    }
                },
                "responses":{
                    "200":{
                        
                        "description": "OK!",
                        "content":{
                            "application/json":{
                                "schema":{
                                    "type":"object",
                                    "items":{
                                        "$ref":"#/components/schemas/pet"
                                    }

                                }
                            }
                        }
                    
                        
                        
                    },
                    "400":{
                        "description": "Erro na requisição"
                        
                    },

                    "404":{
                        "description": "não encontrado"
                    },

                    "500":{
                        "description": "Erro no Servidor"
                    }

                }
            },
            "delete":{
                "description": "Exclui o pet pelo numero do pet",
                "summary": "Exclui o pet pelo numero do pet",
                "tags":["pet"],
                "parameters": [
                    {
                        "name": "nroPet",
                        "in": "path",
                        "description": "numero do pet para Exclui",
                        "required": true
                    }
                ],
                
                "responses":{
                    "200":{
                        
                        "description": "OK!",
                        "content":{
                            "application/json":{
                                "schema":{
                                    "type":"object",
                                    "items":{
                                        "$ref":"#/components/schemas/pet"
                                    }

                                }
                            }
                        }
                    
                        
                        
                    },
                    "400":{
                        "description": "Erro na requisição"
                        
                    },

                    "404":{
                        "description": "não encontrado"
                    },

                    "500":{
                        "description": "Erro no Servidor"
                    }

                }

            }
        },
...
    },
                "components":{
                    "schemas": {
...
                        "pet":{
                            "type": "object",
                            "properties":{
                                "nroPet": {
                                    "type": "number"
                                },
                                "nome": {
                                    "type": "string"
                                },
                                "tipo": {
                                    "type": "string"
                                },
                                 "raça": {
                                    "type":"string"
                                 },
                                "idade": {
                                    "type": "number"
                                },
                                "peso": {
                                    "type":"number"
                                },
                                "tutor":{
                                    "type":"objectId"
                                },
                                "id":{
                                    "type":"string"
                                }
                            }
                        },
...
                    }
                }
}